# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "laravel-api-main",
        "description": "Description de votre API",
        "contact": {
            "name": "Nom du contact",
            "email": "contact@example.com"
        },
        "license": {
            "name": "Licence",
            "url": "https://example.com/license"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Afficher une catégorie spécifique",
                "operationId": "46c5c3954df07301e16ee08eaf957b86",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne la catégorie spécifique",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categorie"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Mettre à jour une catégorie",
                "operationId": "c64176ec9c818ab3457de40af939694c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la catégorie à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategorieRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Catégorie mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categorie"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Supprimer une catégorie",
                "operationId": "7b13d3ed60e6e7337941a54b21084983",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la catégorie à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catégorie supprimée avec succès"
                    }
                }
            }
        },
        "/api/v1/categories/{id}/edit": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Afficher le formulaire d'édition d'une catégorie",
                "operationId": "f35dc20c7fcd3fcf888d18e9b58e607e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la catégorie à éditer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne la catégorie à éditer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categorie"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CategorieRequest": {
                "title": "Categorie Request",
                "description": "Request body for creating or updating a Categorie.",
                "properties": {
                    "name": {
                        "description": "Name of the categorie",
                        "type": "string",
                        "example": "Category Name"
                    },
                    "description": {
                        "description": "Description of the categorie",
                        "type": "string",
                        "example": "Category Description"
                    }
                },
                "type": "object"
            },
            "Categorie": {
                "title": "Categorie",
                "description": "Categorie model",
                "properties": {
                    "id": {
                        "description": "Categorie ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Categorie name",
                        "type": "string"
                    },
                    "description": {
                        "description": "Categorie description",
                        "type": "string"
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Product"
                        }
                    },
                    "created_at": {
                        "description": "Date and time when the categorie was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date and time when the categorie was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Product": {
                "title": "Product",
                "description": "Product model",
                "properties": {
                    "id": {
                        "description": "Product ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Product name",
                        "type": "string"
                    },
                    "price": {
                        "description": "Product price",
                        "type": "number",
                        "format": "float"
                    },
                    "stock": {
                        "description": "Product stock",
                        "type": "integer"
                    },
                    "description": {
                        "description": "Product description",
                        "type": "string"
                    },
                    "image": {
                        "description": "URL of the product image",
                        "type": "string"
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Categorie"
                        }
                    },
                    "created_at": {
                        "description": "Date and time when the product was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date and time when the product was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    }
}